add category

+++ model/cart.go
@@ -4,10 +4,10 @@ import "gorm.io/gorm"
 
 type Cart struct {
 	gorm.Model
-	UserId    uint `gorm:"not null"`
-	ProductId uint `gorm:"not null"`
-	BossId    uint `gorm:"not null"`
-	Num       uint `gorm:"not null"`
-	Max       uint `gorm:"not null"`
+	UserId    uint  `gorm:"not null"`
+	ProductId int64 `gorm:"not null"`
+	BossId    int64 `gorm:"not null"`
+	Num       int64 `gorm:"not null"`
+	Max       int64 `gorm:"not null"`
 	Check     bool
 }
